requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.34.2
httpx==0.27.2
redis==5.0.8
PyYAML==6.0.2
python-dotenv==1.0.1
google-api-python-client==2.143.0
google-auth==2.34.0
mercadopago==2.3.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pytest-httpx==0.30.0
openai>=1.0.0
pytz==2024.2
